+++ src/main/kotlin/model/Car.kt
@@ -2,5 +2,5 @@ package model
 
 data class Car(
     val name: String,
-    var position: Int
+    var position: Int,
 )
